lambdas/date_generation/requirements.txt
aws_lambda_powertools==2.8.0
datefinder==0.7.3
pandas==2.0.1
python_dateutil==2.8.2
spacy==3.5.0
